ClothingSlot去掉原版的四个逐一枚举，而是统一使用ClothingSlot，让模组的衣物槽能正常保存

+++ Survivalcraft/Game/ClothingData.cs
@@ -167,7 +167,7 @@ namespace Game
 		/// <param name="componentClothing"></param>
 		public virtual void OnClotheSet(ComponentClothing componentClothing)
 		{
-			componentClothing.InsulationBySlots[(int)Slot] += Insulation;
+			componentClothing.InsulationBySlots[Slot] += Insulation;
 			componentClothing.SteedMovementSpeedFactor *= SteedMovementSpeedFactor;
 			componentClothing.m_densityModifierApplied += DensityModifier;
 		}
